% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AudioProcessing.R
\name{AudioProcessing}
\alias{AudioProcessing}
\title{AudioProcessing}
\usage{
AudioProcessing(d, rate = 256, preprocess = "none", filtering = "lgf",
  output = "rmax", params)
}
\arguments{
\item{d}{A two-column data.frame with the first column being time and the second being signal}

\item{rate}{Sampling rate for the signal}

\item{preprocess}{Options are 'ma' (Moving average) and 'none'}

\item{filtering}{Options are 'lgf' (local Gaussian Filtering), 'gf' (Gaussian Filtering),
'mgf' (Mixed Gaussian Filtering, not implemented), and 'none'}

\item{output}{Options are 'rmean' (running mean), 'rmedian' (running median),
'apnea', and 'none'}

\item{params}{Nested list of parameters, or path to a yaml config file (see Details)}

\item{thresh_right}{Threshold to filter out signal from the end to remove artifacts like
removing device and the like, that can bias results. There are two options:
1. stoptime: Time (in seconds) from the end when we stop
2. last_window: which is the last observation (index) to keep (is this good?)}
}
\description{
Signal processing the zdevice audio signal
}
\details{
This function allows the modular implementation of different pre-processing, filtering
and summarization steps in processing zdevice audio signal.
There are several choices for each of these
steps already implemented.

\strong{Pre-processing}
\enumerate{
\item "none" : No pre-processing
\item "ma" : Run a moving average on the raw signal
\itemize{
\item "window" : The width of the moving average window in seconds
}
}

\strong{Filtering}
\enumerate{
\item "none": No filtering
\item "gf" : Gaussian filtering
\itemize{
\item "threshold": The threshold value beyond which we consider the filtered values to be signal
\item "robust": Use \code{median} and \code{mad} to scale the data instead of \code{mean} and \code{sd}
}
\item "lgf": Local Gaussian Filtering (Default)
\itemize{
\item "threshold": The threshold value beyond which we consider the filtered values to be signal
\item "window" : The window for computing the local mean and sd for scaling the data, in seconds
\item "robust": \emph{not implemented}
}
}

\strong{Summarizing}
\enumerate{
\item "none" : no summarizing
\item "rmax" : Moving maximum (Default)
\itemize{
\item "window": Window of the moving average in seconds
}
\item "rmean" : Moving average
\itemize{
\item "window": Window of moving average in seconds
}
\item "apnea" : The apnea detection methodology (not implemented)
}

\strong{Specifying parameters:} One way to specify the parameters in this function call is to create
a YAML file of the parameters and read it with \code{yaml::read_yaml}. This would get to a nested list
that could be passed as the \code{params} object. For example::\preformatted{pre:
  window: 1
filt:
  robust: FALSE
  threshold: 3
  roll_window: 30
output:
  window: 1
thresh_right:
  stoptime: 2000

}
}
\examples{

}
