% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrated_data.R
\name{calibrated_data}
\alias{calibrated_data}
\title{Combining sound and motion into data channels at same times}
\usage{
calibrated_data(dat)
}
\arguments{
\item{dat}{A list, output from \code{readData}.}
}
\value{
A single data.frame/tibble containing 5 columns: time, x, y, z, sound
}
\description{
Combining sound and motion into data channels at same times
}
