% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_peaks.R
\name{find_peaks}
\alias{find_peaks}
\title{find_peaks}
\usage{
find_peaks(x, m = 3)
}
\arguments{
\item{x}{A numerical vector}

\item{m}{The window that defines the local maximum (default = 3)}
}
\value{
A vector of indices where the peaks have been identified
}
\description{
This function finds peaks in a time series of data, defined to be a local maxima where
at least \code{m} points on either side are smaller than the deemed peak
}
\details{
This function is written by Stasia Grinberg and is available at https://github.com/stas-g/findPeaks
}
