% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussFilter.R
\name{GaussFilter}
\alias{GaussFilter}
\title{Gaussian Filtering of signal}
\usage{
GaussFilter(x, threshold, robust = FALSE)
}
\arguments{
\item{x}{A numeric vector comprising the signal}

\item{threshold}{The cutoff above which the absolute transformed signal will be considered true signal}

\item{robust}{A logical. Should robust scaling be used (default = FALSE)}
}
\value{

}
\description{
This function takes a signal, calculates the (robust) Z-transform,
takes its absolute value, subtracts the threshold and returns the positive
part of the transformed signal
}
