% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalGaussFilter.R
\name{LocalGaussFilter}
\alias{LocalGaussFilter}
\title{LocalGaussFilter}
\usage{
LocalGaussFilter(x, n, threshold, robust = FALSE)
}
\arguments{
\item{x}{A vector of numerical values}

\item{n}{The width of the window to perform rolling calculations on}

\item{threshold}{The threshold beyond which the normalized data is considered signal}

\item{robust}{(not implemented)}
}
\value{
A new numerical vector of thresholded values
}
\description{
This function takes a signal and applies moving means and SDs to it of a particular
window (\code{n}). It then scales the data by subtracting from each point the average of the
previous values in the window and dividing by their SD. The absolute value is then taken and
a threshold (\code{threshold}) is applied to determine the signal.
}
\details{
A robust option using rolling median and MAD is yet to be implemented
}
\examples{
x <- rnorm(1000000)
y <- LocalGaussFilter(x, n = 100, threshold = 3)

}
