% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_resp_var.R
\name{plot_resp_var}
\alias{plot_resp_var}
\title{Plot breathing variability}
\usage{
plot_resp_var(d)
}
\arguments{
\item{d}{data.frame, output of process_run_v2}
}
\description{
Plot breathing variability
}
