% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_resp_zone.R
\name{plot_resp_zones}
\alias{plot_resp_zones}
\title{Plotting breathing zones}
\usage{
plot_resp_zones(d)
}
\arguments{
\item{d}{data.frame object, output of process_run_v2}
}
\description{
Plotting breathing zones
}
