% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_run.R
\name{process_run}
\alias{process_run}
\title{process_run}
\usage{
process_run(D)
}
\arguments{
\item{D}{A data.frame processed through zdeviceR::calibrated_data, downsampled to 10Hz}
}
\value{
A data.frame with time, Cadence and Respiration rate
}
\description{
Processing run data to get cadence and respiration rate
}
