% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_run_v2.R
\name{process_run_v2}
\alias{process_run_v2}
\title{Process run data from calibrated (10 Hz) data}
\usage{
process_run_v2(d)
}
\arguments{
\item{d}{data.frame, output from \code{calibrated_data}}
}
\value{
data.frame with various run statistics
}
\description{
Process run data from calibrated (10 Hz) data
}
