% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\title{Reading data from zdevice}
\usage{
readData(filename, sound_rate = 256, accel_rate = 10)
}
\arguments{
\item{filename}{A string with the path to the data file}

\item{sound_rate}{Sampling rate for sound data (default = 256)}

\item{accel_rate}{Sampling rate for motion data (default = 10)}
}
\value{
A list with 3 components:
   \enumerate{
       \item timestamp: The timestamp when the recording was started
       \item Sound: A data.frame with time and sound data
       \item Motion: A data.frame with time and motion data (x, y, and z axes)}
}
\description{
Reading data from zdevice
}
\examples{
dat <- readData(system.file('extdata','example_data.txt', package='zdeviceR'))
sound_data <- dat[['Sound']]
motion_data <- dat[['Motion']]
}
