% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSoundMotion.R
\name{readMotion}
\alias{readMotion}
\title{readMotion}
\usage{
readMotion(filename, accel_rate = 10)
}
\arguments{
\item{filename}{Text field giving path to file to be read}

\item{accel_rate}{Rate that motion was sampled, in Hz}
}
\value{
a data.frame with timestamp (Time) and x, y and z direction accelerations
}
\description{
readMotion
}
\examples{
accel_dat <- readMotion(system.file('extdata', 'example_data.txt'), package='zdeviceR')

}
