% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSoundMotion.R
\name{readSound}
\alias{readSound}
\title{readSound}
\usage{
readSound(filename, sound_rate = 256)
}
\arguments{
\item{filename}{Text field giving path to file to be read}

\item{sound_rate}{Rate that sound is sampled, in Hz}
}
\value{
a data.frame with timestamp (Time) and sound intensity (Sound) columns
}
\description{
Reading sound data from zdevice downloads
}
\examples{
sound_dat <- readSound(system.file('extdata', 'example_data.txt'), package='zdeviceR')
}
