% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignalProcessing.R
\name{SignalProcessing}
\alias{SignalProcessing}
\title{Signal processing the zdevice signal}
\usage{
SignalProcessing(d, rate = 256, thresh_right = c(stoptime = 2000,
  last_window = NULL), preprocess = "ma", filtering = "lgf",
  output = "rmax", preParams = list(window = 256L),
  filtParams = list(threshold = 3.5), outputParams = list(window = 256L))
}
\arguments{
\item{d}{A two-column data.frame with the first column being time and the second being signal}

\item{preprocess}{Options are 'ma' (Moving average) and 'none'}

\item{filtering}{Options are 'lgf' (local Gaussian Filtering), 'gf' (Gaussian Filtering),
'mgf' (Mixed Gaussian Filtering, not implemented), and 'none'}

\item{output}{Options are 'rmean' (running mean), 'rmedian' (running median),
'apnea', and 'none'}

\item{preParams}{Parameters for the chosen preprocess method, as a list}

\item{filtParams}{Parameters for the chosen filtering method, as a list}

\item{outputParams}{Parameters for the chosen output method, as a list}
}
\description{
Signal processing the zdevice signal
}
