% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zdeviceR-package.R, R/zzz.R
\docType{package}
\name{zdeviceR}
\alias{zdeviceR}
\alias{zdeviceR-package}
\alias{zdeviceR}
\alias{zdeviceR-package}
\title{zdeviceR}
\description{
Loading and processing data from zdevice

Loading and processing data from zdevice
}
\author{
Abhijit Dasgupta

Abhijit Dasgupta
}
